[FIX] : 🐛 패키지 명 충돌 수정
