+++ app/src/main/java/com/android/youtubeproject/api/model/YoutubeModel.kt
@@ -1,3 +1,3 @@
-package com.android.youtubeproject.api
+package com.android.youtubeproject.api.model
 
 class YoutubeModel(var type:Int, var title:String, var url:String)
