+++ app/src/main/res/layout/fragment_my_video.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".MyVideoFragment.MyVideoFragment">
+    tools:context=".fragment.myvideofragment.MyVideoFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
