+++ app/src/main/res/layout/fragment_video_detail.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".VideoDetailFragment.VideoDetailFragment">
+    tools:context=".fragment.videodetailfragment.VideoDetailFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
